update build: suppress gki mode for KSU
Since KernelSU v3.0.0 the GKI mode is deprecated. In result, you get in your Manager a warning which you can't get rid of.
Due to SuSFS and/or HymoFS we need to build it as GKI image. So we suppress the GKI and make it as fake LKM mode.
When you don't want this behaviour, simply run the ./setup_buildscript.sh with the --disableFakeLKM flag.